+++ spec/cached_defaults/dummy_models_controller_spec.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe CachedDefaults::DummyModelsController, type: :controller do
+  before { Rails.cache.clear }
+
+  include_examples 'simple crud for show'
+
+  describe 'GET #index' do
+    before { create_list(:dummy_model, 2) }
+
+    it 'returns all records' do
+      get :index
+      expect(response.parsed_body.length).to eq(2)
+    end
+  end
+end